==> spark/15_Load_DW.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# matplotlib 한글 꺠짐 방지
# apt-get update
# apt-get install fonts-nanum* 
# apt-get install fontconfig
# fc-cache -fv  # font 캐시 날리기
# rm -rf /home/hy1/.cache/matplotlib/  #matplotliob 폰트 캐시 날리기



from datetime import date, datetime
from pyspark.sql.types import *
from pyspark.sql.functions import *
from pyspark.sql import *
import datetime as dt
import pandas as pd
import matplotlib.pyplot as plt #그래프 패키지 모듈 등록
%matplotlib inline 


def cal_std_day(befor_day):   
    x = dt.datetime.now() - dt.timedelta(befor_day)
    year = x.year
    month = x.month if x.month >= 10 else '0'+ str(x.month)
    day = x.day if x.day >= 10 else '0'+ str(x.day)  
    return str(year)+ '-' +str(month)+ '-' +str(day)

import platform

from matplotlib import font_manager, rc
plt.rcParams['axes.unicode_minus'] = False

if platform.system() == 'Darwin':  # 맥OS 
    rc('font', family='AppleGothic')
elif platform.system() == 'Windows':  # 윈도우
    path = "c:/Windows/Fonts/malgun.ttf"
    font_name = font_manager.FontProperties(fname=path).get_name()
    rc('font', family=font_name)
else:
    rc('font', family='NanumGothic')

## Load - DataWareHouse

## JDBC DICT

JDBC = {
    'url':'jdbc:oracle:thin:@decorona_high?TNS_ADMIN=/home/big/study/db/Wallet_DECORONA'
    , 'props':{
        'user':'dw_de'
        ,'password':'123qwe!@#QWE'
    }
}

### 1. LOC 테이블 저장

area = spark.read.csv('/corona_data/loc/sido_area.csv', encoding='CP949', header=True)
popu = spark.read.csv('/corona_data/loc/sido_population.csv', encoding='CP949', header=True)


# loc 데이터 저장
area_pop = area.join(popu, on='loc')
area_pop = area_pop.select(col('loc').alias('LOC')
                          ,col('area').alias('AREA')
                          ,col('total').alias('POPULATION'))
area_pop.show(5)
area_pop.write.jdbc(url=JDBC['url'], table='LOC', mode='append', properties=JDBC['props'])

### 2. CORONA_PATIENT 테이블 저장

#  코로나 감염자 데이터
path = '/corona_data/patient'
co_patient_json = spark.read.json(path, encoding='UTF-8')
# co_patient_json.show()

data = []
for r1 in co_patient_json.select('items').toLocalIterator():
    if not r1.items:
        continue
    for r2 in r1.items:
        data.append(r2)

patient_data = spark.createDataFrame(data)
patient_data.show(3)

co_patients = patient_data.select(
    patient_data.gubun.alias('LOC')
    ,patient_data.deathCnt.alias('DEATH_CNT')
     ,patient_data.defCnt.alias('DEF_CNT')
     ,patient_data.localOccCnt.alias('LOC_OCC_CNT')
     ,patient_data.qurRate.alias('QUR_RATE')
     ,patient_data.stdDay.alias('STD_DAY')
).where(~(col('LOC').isin(['합계','검역'])) ).distinct()
co_patients.printSchema()

co_patients.write.jdbc(url=JDBC['url'], table='CORONA_PATIENTS', mode='append', properties=JDBC['props'])

### 3.  CORONA_VACCINE 테이블 저장

file_name = '/corona_data/vaccine/'
vaccine = spark.read.json(file_name, multiLine=True)
# vaccine.printSchema()
# vaccine.show()
vaccine.show()


vaccine.select('meta').collect()

data = []
for r1 in vaccine.select(vaccine.data, vaccine.meta.std_day).toLocalIterator():
    for r2 in r1.data:
        # row 객체를 만들기 위해 dict 생성
        # ** : 파이썬 압축해제 키워드 
        # **kwargs   => 여러 쌍의 키워드 args 가 넘어오면 받아서 dictionary로 반환
        # fnc(**dict) => dict에 있는 key-value 값들이 여러쌍의 kwargs  형태로 함수에 전달
        temp = r2.asDict()
        temp['std_day'] = r1['meta.std_day']
        data.append(Row(**temp))
        
vaccine_data = spark.createDataFrame(data)
vaccine_data.show()

# column 을 row 로 변환 : stack
# row 를 colum 으로 변환 : pivot
pd_vaccine = vaccine_data.to_pandas_on_spark()
pd_vaccine = pd_vaccine.set_index(['loc','std_day'])
pd_vaccine = pd_vaccine.stack()
# 시리즈나 데이터프레임으로 반환돠는데, 이 때 아래 사용
pd_vaccine = pd_vaccine.to_dataframe('V_CNT') # 0이 컬럼명에 들어가 있는 거 확인  V_CNT로 변경
pd_vaccine

pd_vaccine = pd_vaccine.reset_index()
pd_vaccine= pd_vaccine.rename(columns={'level_2':'V_TH'})
vaccine_data.drop('level_0')
vaccine_data.drop('index')
vaccine_data.show()

vaccine_data = pd_vaccine.to_spark()
vaccine_data.show()

vaccine_data.write.jdbc(url=JDBC['url'], table='CORONA_VACCINE', mode='append', properties=JDBC['props'])

### 4. LOC_FACILITY_CNT	테이블 저장
