==> spark/16_Load_DM.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# 데이터마트용 계정 생성
# sqlDeveloper에서 진행
# create user dm_admin IDENTIFIED BY "123qwe!@#QWE";
# grant dwrole to dm_admin;
# grant unlimited tablespace to dm_admin;


from datetime import date, datetime
from pyspark.sql.types import *
from pyspark.sql.functions import *
import datetime as dt
import pandas as pd
import matplotlib.pyplot as plt #그래프 패키지 모듈 등록
%matplotlib inline 


import platform

from matplotlib import font_manager, rc
plt.rcParams['axes.unicode_minus'] = False

if platform.system() == 'Darwin':  # 맥OS 
    rc('font', family='AppleGothic')
elif platform.system() == 'Windows':  # 윈도우
    path = "c:/Windows/Fonts/malgun.ttf"
    font_name = font_manager.FontProperties(fname=path).get_name()
    rc('font', family=font_name)
else:
    rc('font', family='NanumGothic')

def cal_std_day(befor_day):   
    x = dt.datetime.now() - dt.timedelta(befor_day)
    year = x.year
    month = x.month if x.month >= 10 else '0'+ str(x.month)
    day = x.day if x.day >= 10 else '0'+ str(x.day)  
    return str(year)+ '-' +str(month)+ '-' +str(day)

### JDBC

JDBC_DW= {
    'url':'jdbc:oracle:thin:@decorona_high?TNS_ADMIN=/home/big/study/db/Wallet_DECORONA'
    , 'props':{
        'user':'dw_de'
        ,'password':'123qwe!@#QWE'
    }
}

JDBC_DM= {
    'url':'jdbc:oracle:thin:@decorona_high?TNS_ADMIN=/home/big/study/db/Wallet_DECORONA'
    , 'props':{
        'user':'admin'
        ,'password':'123qwe!@#QWE'
    }
}

## Load DataMart

### 1. 지역별 단위면적(km)당 인구수와 코로나 확진자 수 데이터 저장

popu = spark.read.jdbc(url=JDBC_DW['url'],table='loc',properties=JDBC_DW['props'])
popu.show(3)

patients = spark.read.jdbc(url=JDBC_DW['url'],table='CORONA_PATIENTS',properties=JDBC_DW['props'])
patients.show(3)

pop_patients = popu.join(patients, on = 'loc') \
                        .select('loc'
                               ,ceil(col('population')/col('area')).alias('POPU_DENSITY')
                               ,'qur_rate'
                               ,'std_day') \
                        .orderBy(col('std_day'))

pop_patients.show()

pop_patients.write.jdbc(url=JDBC_DM['url'], table='CO_POPU_DENSITY',mode='append',properties=JDBC_DM['props'])

### 2.  백신접종 완료자와 코로나 확진자 수 데이터 저장

vaccine = spark.read.jdbc(url=JDBC_DW['url'],table='CORONA_VACCINE',properties=JDBC_DW['props'])
vaccine.show(3)

pd_vaccine = vaccine.to_pandas_on_spark()
pd_vaccine = pd_vaccine.pivot_table(index=['LOC','STD_DAY'],columns='V_TH',values='V_CNT')
pd_vaccine
pd_vaccine = pd_vaccine.reset_index()
pd_vaccine
tmp = pd_vaccine.to_spark()

vac_rate = tmp.join(popu, on='LOC') \
                    .join(patients, on='LOC') \
                        .select('loc', patients.STD_DAY
                               ,ceil(col('v1')/col('population')*100).alias('V1TH_RATE')
                               ,ceil(col('v2')/col('population')*100).alias('V2TH_RATE')
                               ,ceil(col('v3')/col('population')*100).alias('V3TH_RATE')
                               ,ceil(col('v4')/col('population')*100).alias('V4TH_RATE')
                               ,'qur_rate')

vac_rate.show(3)

vac_rate.write.jdbc(url=JDBC_DM['url'], table='CO_VACCINE',mode='append',properties=JDBC_DM['props'])

### 3. 다중이용시설과 코로나 확진자 수의 상관관계

fac = spark.read.csv('/corona_data/loc/전국다중이용시설.csv', encoding='CP949', header=True)
fac.show(5)
popu = spark.read.jdbc(url=JDBC_DW['url'],table='loc',properties=JDBC_DW['props'])
popu.show()
patients = spark.read.jdbc(url=JDBC_DW['url'],table='corona_patients',properties=JDBC_DW['props'])
patients.show(5)

fac_cnt = fac.groupBy(fac.광역.alias('loc')).agg(count('*').alias('fac_cnt'))
fac_cnt.show()
tmp = popu.join(fac_cnt, on='LOC')
tmp.show(5)
fac_popu = tmp.select('LOC', ceil(col('FACILITY_CNT')/col('POPULATION')*100000).alias('FAC_POPU'))
fac_popu.show()
co_fac = fac_popu.join(patients, on='LOC') \
                    .select('LOC'
                            , 'FAC_POPU'
                           , 'QUR_RATE'
                           , 'STD_DAY')
co_fac.show(10)

co_fac.write.jdbc(url=JDBC_DM['url'], table='CO_FACILITY',mode='append',properties=JDBC_DM['props'])

### 4. 요일별 코로나 확진자 수를 구해보자

patients = spark.read.jdbc(url=JDBC_DW['url'],table='CORONA_PATIENTS',properties=JDBC_DW['props'])
patients.show(10)

day_patients = patients.withColumn('day_of_week', dayofweek(col('STD_DAY')))
pd_day_patients = day_patients.to_pandas_on_spark()
pd_day_patients = pd_day_patients.pivot_table(index=['STD_DAY'],columns='day_of_week',values='LOC_OCC_CNT')
pd_day_patients
pd_day_patients = pd_day_patients.reset_index()


# day_patients = day_patients.select('day_of_week', 'STD_DAY').show(10)
# pd_day_patients = day_patients.to_pandas_on_spark()
# pd_day_patients
